% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktimr.R
\name{promptTimer}
\alias{promptTimer}
\title{Prompt to repeat \code{workTimer}}
\usage{
promptTimer(minutes = 25, lockScreen = TRUE)
}
\arguments{
\item{minutes}{numeric}

\item{lockScreen}{boolean lock the screen when countdown ends}
}
\description{
Keep prompting to start another batch of work until quit.
}
