% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktimr.R
\name{promptTimer}
\alias{promptTimer}
\title{Prompt to repeat \code{workTimer}}
\usage{
promptTimer(minutes = 25, outcome = "lock", con = "",
  timerMessage = "time until explosion: ", finalMessage = "IT'S OVER!")
}
\arguments{
\item{minutes}{numeric}

\item{outcome}{character naming one of "lock", "alarm", or "nothing"}

\item{con}{character naming output text file}

\item{timerMessage}{character display before countdown timer}

\item{finalMessage}{character display when countdown expired}
}
\description{
Keep prompting to start another batch of work until quit.
}
