% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktimr.R
\name{promptTimer}
\alias{promptTimer}
\title{Prompt to repeat \code{workTimer}}
\usage{
promptTimer(explosion, minutes = 25)
}
\arguments{
\item{explosion}{file path to image file}

\item{minutes}{numeric}
}
\description{
Keep prompting to start another batch of work until quit.
}
