% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktimr.R
\name{workTimer}
\alias{workTimer}
\title{Countdown until stop work}
\usage{
workTimer(minutes = 25, lockScreen = TRUE)
}
\arguments{
\item{minutes}{numeric}

\item{lockScreen}{boolean lock the screen when countdown ends}
}
\description{
Countdown \code{minutes} then lock screen.
}
