% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktimr.R
\name{workTimer}
\alias{workTimer}
\title{Countdown until stop work}
\usage{
workTimer(explosion, minutes = 25)
}
\arguments{
\item{explosion}{file path to image file}

\item{minutes}{numeric}
}
\description{
Countdown \code{minutes} then show \code{explosion} image
fullscreen.
}
